g2pe/requirements.txt
numpy
opencv-python-headless
Pillow
requests
streamlit 
